drfecommerce/requirement.txt
asgiref==3.11.0
Django==6.0
django-js-asset==3.1.2
django-mptt==0.18.0
djangorestframework==3.16.1
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.1
pytest-django==4.11.1
python-dotenv==1.2.1
sqlparse==0.5.4
